+++ src/test/java/calculator/ParserTest.java
@@ -2,7 +2,7 @@ package calculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import calculator.Parser;
+import calculator.model.Parser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
